Reduce and organise some views/partials
Some views were not in use. Some partials were redundant.

Also cleanup some locales related to surveys that we don't use in this service (that was a leftover from another service).

The aim is to reduce the amount of views and partials that we need to migrate to the ones in use and do not waste time with things that are not in use.
