+++ app/controllers/entrypoint_controller.rb
@@ -1,10 +1,6 @@
 class EntrypointController < ApplicationController
   def v1; end
 
-  def v2; end
-
-  def v3; end
-
   def what_is_needed; end
 
   def how_long; end
