--- app/controllers/home_controller.rb
@@ -1,3 +1,0 @@
-class HomeController < ApplicationController
-  def miam_exemptions; end
-end
