--- app/views/entrypoint/shared/_start_application.html.erb
@@ -1,3 +1,0 @@
-<div class="xform-group form-submit">
-  <%= link_to t('.start'), entrypoint_what_is_needed_path, class: 'button button-start', role: 'button' %>
-</div>
