+++ app/views/entrypoint/v1.en.html.erb
@@ -42,7 +42,7 @@
           Read the <a href="https://helpwithchildarrangements.service.justice.gov.uk" rel="external" target="_blank">child arrangements guide</a> to see if there’s a more suitable option than going to court.
           </li>
           <li><h3 class="heading-small">Check if you have a valid reason for not attending a Mediation Information and Assessment Meeting (<abbr title="Mediation Information and Assessment Meeting">MIAM</abbr>)</h3>
-          A <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr> is an introductory meeting that explains how <a href="https://helpwithchildarrangements.service.justice.gov.uk/professional-mediation" rel="external" target="_blank">mediation</a> can help resolve your dispute. You’re legally required to attend one if you want to go to court, unless <a href="/miam_exemptions" target="_blank">you’re exempt</a>.
+          A <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr> is an introductory meeting that explains how <a href="https://helpwithchildarrangements.service.justice.gov.uk/professional-mediation" rel="external" target="_blank">mediation</a> can help resolve your dispute. You’re legally required to attend one if you want to go to court, unless <a href="/about/miam_exemptions" target="_blank">you’re exempt</a>.
           </li>
           <li><h3 class="heading-small">Find a mediator and book a <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr></h3>
           If you have to attend a <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr> you’ll need to <a href="https://www.familymediationcouncil.org.uk/find-local-mediator/" rel="external" target="_blank">find a mediator</a> to book one.
@@ -63,6 +63,11 @@
     </section>
     </div>
 
-    <%= render partial: 'entrypoint/shared/start_application' %>
+    <%= link_to entrypoint_what_is_needed_path, class: 'govuk-button govuk-button--start govuk-!-margin-top-3', data: { module: 'govuk-button' }, role: 'button', draggable: false do %>
+      Start application
+      <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" aria-hidden="true" focusable="false">
+        <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z" />
+      </svg>
+    <% end %>
   </div>
 </div>
