--- app/views/entrypoint/v2.en.html.erb
@@ -1,9 +1,0 @@
-<% title 'Entrypoint version 2' %>
-
-<div class="grid-row">
-  <div class="column-two-thirds">
-    <h1 class="heading-xlarge gv-u-heading-xxlarge">Entrypoint version 2</h1>
-
-    <%= render partial: 'entrypoint/shared/start_application' %>
-  </div>
-</div>
