--- app/views/entrypoint/v3.en.html.erb
@@ -1,9 +1,0 @@
-<% title 'Entrypoint version 3' %>
-
-<div class="grid-row">
-  <div class="column-two-thirds">
-    <h1 class="heading-xlarge gv-u-heading-xxlarge">Entrypoint version 3</h1>
-
-    <%= render partial: 'entrypoint/shared/start_application' %>
-  </div>
-</div>
