+++ app/views/layouts/_step_header.html.erb
