+++ app/views/steps/miam/acknowledgement/edit.en.html.erb
@@ -6,7 +6,7 @@
 
     <div class="grid_content_header">
       <h1 class="heading-xlarge gv-u-heading-xxlarge">Attending a Mediation Information and Assessment Meeting (MIAM)</h1>
-      <p class="lede gv-u-text-lede">Before completing this application you’re legally required to attend a Mediation Information and Assessment Meeting (<abbr title="Mediation Information and Assessment Meeting">MIAM</abbr>), unless <a href="/miam_exemptions" target="_blank">you’re exempt</a>.</p>
+      <p class="lede gv-u-text-lede">Before completing this application you’re legally required to attend a Mediation Information and Assessment Meeting (<abbr title="Mediation Information and Assessment Meeting">MIAM</abbr>), unless <a href="/about/miam_exemptions" target="_blank">you’re exempt</a>.</p>
     </div>
 
     <div class="govuk-govspeak gv-s-prose">
@@ -34,7 +34,7 @@
 
       <h2 class="gv-u-heading-xlarge">When you don’t have to attend a <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr></h2>
       <p>You don’t have to attend a <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr> if you have a valid reason. For example you or the children are at risk of harm.</p>
-      <p><a href="/miam_exemptions" target="_blank">List of valid reasons for not attending a <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr></a></p>
+      <p><a href="/about/miam_exemptions" target="_blank">List of valid reasons for not attending a <abbr title="Mediation Information and Assessment Meeting">MIAM</abbr></a></p>
     </div>
 
     <%= step_form @form_object do |f| %>
