+++ app/views/steps/miam/certification_expired_info/show.html.erb
@@ -13,7 +13,7 @@
 
     <div class=" govuk-govspeak gv-s-prose">
       <p>
-        <a href="/miam_exemptions" target="_blank">Check if you have
+        <a href="/about/miam_exemptions" target="_blank">Check if you have
           a valid reason</a> before you continue with your application.</p>
 
       <h2 class="heading-medium">If you don’t have a valid reason</h2>
